requirements.txt
pytz==2018.9  # https://github.com/stub42/pytz
redis>=2.10.6, < 3  # pyup: < 3 # https://github.com/antirez/redis
celery==4.2.1  # pyup: < 5.0  # https://github.com/celery/celery
django==2.0.13  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
djangorestframework==3.9.1  # https://github.com/encode/django-rest-framework
psycopg2-binary==2.7.7 
django-filter==2.1.0 